+++ CHANGELOG.md
@@ -19,7 +19,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
-## [0.27.2] - 2024-07-08
+## [0.29.0] - 2024-07-12
+
+### Added
+
+- add validator to base model that verifies computed fields can be set but not altered
+- new class hierarchy for identifiers: ExtractedIdentifier and MergedIdentifier
+
+### Changes
+
+- improve typing for methods using `Self`
+- make local type variables private
+- use json instead of pickle to calculate checksum of models
+- replace `set_identifiers` validator with computed fields on each extracted model
+
+### Removed
+
+- removed custom stringify method on base entities that included the `identifier` field
+
+### Fixed
+
+- fix typing for `__eq__` arguments
+
+## [0.28.0] - 2024-07-08
 
 ### Added